chore: Rename sidebar item injectable files
