+++ packages/core/src/renderer/components/cluster/cluster-overview-sidebar-item.injectable.tsx
