+++ packages/core/src/renderer/components/config/config-sidebar-item.injectable.tsx
