+++ packages/core/src/renderer/components/custom-resources/custom-resource-definitions-sidebar-item.injectable.ts
@@ -7,7 +7,7 @@ import { getInjectable } from "@ogre-tools/injectable";
 import customResourceDefinitionsRouteInjectable from "../../../common/front-end-routing/routes/cluster/custom-resources/custom-resource-definitions.injectable";
 import navigateToCustomResourcesInjectable from "../../../common/front-end-routing/routes/cluster/custom-resources/navigate-to-custom-resources.injectable";
 import routeIsActiveInjectable from "../../routes/route-is-active.injectable";
-import customResourcesSidebarItemInjectable from "./custom-resources-sidebar-items.injectable";
+import customResourcesSidebarItemInjectable from "./custom-resources-sidebar-item.injectable";
 
 const customResourceDefinitionsSidebarItemInjectable = getInjectable({
   id: "custom-resource-definitions-sidebar-item",
