+++ packages/core/src/renderer/components/custom-resources/custom-resources-sidebar-item.injectable.tsx
