+++ packages/core/src/renderer/components/events/events-sidebar-item.injectable.tsx
