+++ packages/core/src/renderer/components/helm/helm-sidebar-item.injectable.tsx
