+++ packages/core/src/renderer/components/namespaces/namespaces-sidebar-item.injectable.tsx
