+++ packages/core/src/renderer/components/network/network-sidebar-item.injectable.tsx
