+++ packages/core/src/renderer/components/nodes/nodes-sidebar-item.injectable.tsx
