+++ packages/core/src/renderer/components/storage-classes/storage-classes-sidebar-item.injectable.tsx
@@ -5,7 +5,7 @@
 import { getInjectable } from "@ogre-tools/injectable";
 
 import storageClassesRouteInjectable from "../../../common/front-end-routing/routes/cluster/storage/storage-classes/storage-classes-route.injectable";
-import storageSidebarItemInjectable from "../storage/storage-sidebar-items.injectable";
+import storageSidebarItemInjectable from "../storage/storage-sidebar-item.injectable";
 import { sidebarItemInjectionToken } from "@k8slens/cluster-sidebar";
 import routeIsActiveInjectable from "../../routes/route-is-active.injectable";
 import navigateToStorageClassesInjectable from "../../../common/front-end-routing/routes/cluster/storage/storage-classes/navigate-to-storage-classes.injectable";
