+++ packages/core/src/renderer/components/storage/storage-sidebar-item.injectable.tsx
