+++ packages/core/src/renderer/components/user-management/user-management-sidebar-item.injectable.tsx
