+++ packages/core/src/renderer/components/workloads/workloads-sidebar-item.injectable.tsx
